add localpkg format stub
